+++ resources/views/default/column/filter/control.blade.php
@@ -3,9 +3,7 @@
         {{ trans('sleeping_owl::lang.table.filters.control') }}
     </button>
 
-    <div class="btn-group">
-        <button type="button" class="btn btn-sm btn-danger" id="filters-cancel">
-            <i class="fas fa-times"></i>
-        </button>
-    </div>
+    <button type="button" class="btn btn-sm btn-danger" id="filters-cancel" data-toggle="tooltip" title="{{ trans('sleeping_owl::lang.button.clear') . ' ' . trans('sleeping_owl::lang.table.all')}}">
+      <i class="fas fa-times"></i>
+    </button>
 </div>